fix(indexeddb): deduplicate DB helpers, add persist error handling, fix plural
- Extract shared indexeddb helpers (openDB, getAll, batchPut, del, clear, count)
  into src/services/indexeddb.ts to eliminate duplication in history.ts and
  history-indexeddb.ts (net -53 lines)
- Add try/catch with console.error to persist() in useHistory so unhandled
  IDBRejectionErrors can't silently crash
- Replace addToIndex-loop with rebuildIndex() in initial load
- useCallback persist deps: [] instead of [persist] (circular ref)
- Fix plural: "result" → "results" in HistoryList search count
